requirements.txt
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.3.0
pytest>=7.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
